ai-sample/requirements.txt
Flask
requests
python-dotenv
prometheus-client
psutil
confluent-kafka